useScrollToItem - change implementation
